requirements.txt
sentence-transformers
typing_extensions
langchain
langchain-community
huggingface_hub
transformers
accelerate
bitsandbytes
openpyxl
datasets
gdown
google-generativeai
pydantic
jsonref
flask
torch
faiss-gpu